+++ test/models/sqlserver/trigger.rb
@@ -1,0 +1,7 @@
+class SSTestTrigger < ActiveRecord::Base
+  self.table_name = 'sst_table_with_trigger'
+end
+
+class SSTestTriggerUuid < ActiveRecord::Base
+  self.table_name = 'sst_table_with_uuid_trigger'
+end