+++ test/models/sqlserver/trigger_history.rb
@@ -1,0 +1,3 @@
+class SSTestTriggerHistory < ActiveRecord::Base
+  self.table_name = 'sst_table_with_trigger_history'
+end